refactor(room): extract useRoomIdentity from useRoomScreenState (C11)
New hook useRoomIdentity wraps getActorIdentity in a useMemo and
owns the delegation validity warning effect. useRoomScreenState
now calls useRoomIdentity instead of inlining the computation.
